Refactor renderer processing into helper
